Renamed mapIcons folder as markers folder
