+++ src/components/map/layers/edit/EditNodeLayer.tsx
@@ -12,7 +12,7 @@ import EventManager from '~/util/EventManager';
 import NodeLocationType from '~/types/NodeLocationType';
 import { NodeStore } from '~/stores/nodeStore';
 import { MapStore, MapFilter } from '~/stores/mapStore';
-import NodeMarker from '../mapIcons/NodeMarker';
+import NodeMarker from '../markers/NodeMarker';
 import { LeafletContext } from '../../Map';
 import ArrowDecorator from '../ArrowDecorator';
