+++ src/components/map/layers/edit/UpsertRoutePathLayer.tsx
@@ -11,8 +11,8 @@ import { MapStore, MapFilter } from '~/stores/mapStore';
 import { ToolbarStore } from '~/stores/toolbarStore';
 import ToolbarTool from '~/enums/toolbarTool';
 import EventManager from '~/util/EventManager';
-import NodeMarker from '../mapIcons/NodeMarker';
-import Marker from '../mapIcons/Marker';
+import NodeMarker from '../markers/NodeMarker';
+import Marker from '../markers/Marker';
 import ArrowDecorator from '../ArrowDecorator';
 import RoutePathNeighborLinkLayer from './RoutePathNeighborLinkLayer';
 import RoutePathCopySeqmentLayer from './routePathCopySeqmentLayer';
