+++ src/components/map/layers/markers/NodeMarker.tsx
