+++ src/components/map/layers/markers/nodeMarker.scss
